Refactoring the file Patient.tsx into multiple files
